% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPlot.R
\name{boxPlot}
\alias{boxPlot}
\title{Wine points and prices box plot}
\usage{
boxPlot(input_country)
}
\arguments{
\item{input_country}{A character of country of wine origin.}
}
\value{
A \code{ggplot} object. The plot of wine prices and points. If the \code{input_country} is in the dataset, a box plot of wine prices and points for that country is generated. If the \code{input_country} is not in the dataset, or if an incorrect country is input, the box plot will be generated with all wine prices and points in the dataset.
}
\description{
The box plot shows the relationship between wine prices and points in different countries.
}
\examples{
boxPlot("US")
boxPlot("world")

}
