% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myapp.R
\name{myapp}
\alias{myapp}
\title{Shiny App about wine ratings dataset}
\usage{
myapp()
}
\value{
A object about shiny app.
}
\description{
The shiny app can be launched from the \code{wine} package and contains the point plot, column plot, box plot and wordcloud related to the wine dataset.
}
