% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_price.R
\name{point_price}
\alias{point_price}
\title{Wine points and prices}
\usage{
point_price(w)
}
\arguments{
\item{w}{A character of country of wine origin.}
}
\value{
A tibble object or a sentence. If the \code{input_country} is in the dataset, it will return a tibble includes mean point, mean price and total number of wines from different country. If the \code{input_country} is not in the dataset, or if an incorrect country is input, it will prompt you with a sentence.
}
\description{
Showing average points and average prices of wines from different countries.
}
\examples{
point_price("Australia")


}
