% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine points, price and description}
\format{
A data frame contains four columns.
\describe{
\item{country}{A list of country of origin.}
\item{points}{A list of points WineEnthusiast rated the wine on a scale of 1-100 (though they say they only post reviews for wines that score >=80).}
\item{price}{A list of the cost for a bottle of the wine.}
\item{description}{A list of flavors and taste profile as written by reviewer.}
}
}
\source{
\url{https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2019/2019-05-28/winemag-data-130k-v2.csv}
}
\usage{
wine
}
\description{
This dataset from Kaggle, It contains the price of the wine, where it originates from, a written description from a sommelier, and a 80–100 point rating.
}
\keyword{datasets}
