% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud.R
\name{wordcloud}
\alias{wordcloud}
\title{Wine wordcloud plot}
\usage{
wordcloud(input_country)
}
\arguments{
\item{input_country}{A character of country of wine origin.}
}
\value{
A \code{wordcloud2} object. The plot of wine descriptions. If the \code{input_country} is in the dataset, a wordcloud plot of wine descriptions for that country is generated. If the \code{input_country} is not in the dataset, or if an incorrect country is input, the wordcloud plot will be generated with all wine descriptions in the dataset.
}
\description{
The wordcloud plot extracts key words from the \code{description} variable to visualize flavors and taste profiles as written by the reviewer.
}
\examples{
wordcloud("US")
wordcloud("world")

}
